refactor(frontend): relocate Person SIN application state machine types to a non-declaration file
